requirements.txt
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.8
Django==5.0.4
django-allauth==0.63.3
django-appconf==1.0.6
django-crispy-forms==2.1
django-debug-toolbar==4.4.5
django-environ==0.11.2
django-extensions==3.2.3
django-livereload-server==0.5.1
django-sslserver==0.22
django-unfold==0.30.0
django_compressor==4.5
Faker==24.9.0
idna==3.7
MarkupSafe==2.1.5
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
rcssmin==1.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
stripe==10.1.0
tornado==6.4.1
typing_extensions==4.12.2
urllib3==2.2.2
watchgod==0.8.2
Werkzeug==3.0.3
